- changed db-nil-p
this function was transforming "no" and even 0.0 (floating point number) to nil.

That behaivour was incorrect for  numbers and probably not desiderable
for the  string no  (in fact  poll with "No"  option was  not rendered
beacause of this fact).

Removed conversion  for non- integer numbers  and prevented conversion
from some  kind of  non empty string  (like "nil" or  "no") via  a key
argument.